requirements.txt
absl-py==1.0.0
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.9
cycler==0.11.0
dataclasses==0.6
ffmpeg==1.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.2
kiwisolver==1.3.1
Markdown==3.3.6
matplotlib==3.3.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.5.4
seaborn==0.11.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
